요청 헤더의 토큰을 처리해 컨트롤러 메서드의 파라미터로 전달

See merge request s12-webmobile1-sub1/S12P11E204!29

+++ backend/butter/src/main/java/com/ssafy/butter/domain/crew/service/CrewServiceImpl.java
@@ -1,5 +1,6 @@
 package com.ssafy.butter.domain.crew.service;
 
+import com.ssafy.butter.auth.dto.AuthInfoDTO;
 import com.ssafy.butter.domain.crew.dto.request.CrewFollowRequestDTO;
 import com.ssafy.butter.domain.crew.dto.request.CrewListRequestDTO;
 import com.ssafy.butter.domain.crew.dto.request.CrewMemberRequestDTO;
@@ -36,11 +37,12 @@ public class CrewServiceImpl implements CrewService {
 
     /**
      * 크루 생성 요청 DTO를 받아 크루를 생성 및 DB에 저장 후 크루 응답 DTO를 반환한다.
+     * @param currentUser 현재 로그인한 유저 정보
      * @param crewSaveRequestDTO 크루 생성 요청 정보를 담은 DTO
      * @return 크루 생성 응답 정보를 담은 DTO
      */
     @Override
-    public CrewResponseDTO createCrew(CrewSaveRequestDTO crewSaveRequestDTO) {
+    public CrewResponseDTO createCrew(AuthInfoDTO currentUser, CrewSaveRequestDTO crewSaveRequestDTO) {
         String imageUrl = null;
         if (crewSaveRequestDTO.image() != null) {
             imageUrl = s3ImageUploader.uploadImage(crewSaveRequestDTO.image());
@@ -53,20 +55,35 @@ public class CrewServiceImpl implements CrewService {
                 .imageUrl(imageUrl)
                 .portfolioVideoUrl(portfolioVideoUrl)
                 .build();
-        return CrewResponseDTO.fromEntity(crewRepository.save(crew));
+        Crew savedCrew = crewRepository.save(crew);
+
+        crewMemberRepository.save(CrewMember.builder()
+                .crew(savedCrew)
+                .member(memberService.findById(currentUser.id()))
+                .build());
+
+        return CrewResponseDTO.fromEntity(savedCrew);
     }
 
     /**
      * 크루 ID와 멤버 ID가 담긴 DTO를 받아 크루 멤버를 DB에 저장한다.
+     * @param currentUser 현재 로그인한 유저 정보
      * @param crewMemberRequestDTO 크루와 크루 멤버 정보를 담은 DTO
      */
     @Override
-    public void createCrewMember(CrewMemberRequestDTO crewMemberRequestDTO) {
+    public void createCrewMember(AuthInfoDTO currentUser, CrewMemberRequestDTO crewMemberRequestDTO) {
+        Member currentMember = memberService.findById(currentUser.id());
         Crew crew = crewRepository.findById(crewMemberRequestDTO.crewId()).orElseThrow();
+        CrewMember currentCrewMember = crewMemberRepository.findByCrewAndMember(crew, currentMember).orElseThrow();
+        if (!currentCrewMember.getIsCrewAdmin()) {
+            throw new IllegalArgumentException("Current user is not crew admin");
+        }
+
         Member member = memberService.findById(crewMemberRequestDTO.memberId());
         crewMemberRepository.findByCrewAndMember(crew, member).ifPresent(crewMember -> {
             throw new IllegalArgumentException("Crew member already exists");
         });
+
         CrewMember crewMember = CrewMember.builder()
                 .crew(crew)
                 .member(member)
@@ -76,12 +93,18 @@ public class CrewServiceImpl implements CrewService {
 
     /**
      * 크루 ID와 멤버 ID를 받아 크루 멤버를 DB에서 삭제한다.
-     * @param crewId 크루를 가리키는 ID
-     * @param memberId 멤버를 가리키는 ID
+     * @param currentUser 현재 로그인한 유저 정보
+     * @param crewId      크루를 가리키는 ID
+     * @param memberId    멤버를 가리키는 ID
      */
     @Override
-    public void deleteCrewMember(Long crewId, Long memberId) {
+    public void deleteCrewMember(AuthInfoDTO currentUser, Long crewId, Long memberId) {
+        Member currentMember = memberService.findById(currentUser.id());
         Crew crew = crewRepository.findById(crewId).orElseThrow();
+        CrewMember currentCrewMember = crewMemberRepository.findByCrewAndMember(crew, currentMember).orElseThrow();
+        if (!currentCrewMember.getIsCrewAdmin()) {
+            throw new IllegalArgumentException("Current user is not crew admin");
+        }
         Member member = memberService.findById(memberId);
         CrewMember crewMember = crewMemberRepository.findByCrewAndMember(crew, member).orElseThrow();
         crewMemberRepository.delete(crewMember);
@@ -119,13 +142,19 @@ public class CrewServiceImpl implements CrewService {
 
     /**
      * 수정하려는 크루의 ID와 수정할 내용을 받아 DB에 반영하고 수정 결과를 반환한다.
-     * @param id 수정하려는 크루의 ID
+     * @param currentUser 현재 로그인한 유저 정보
+     * @param id                 수정하려는 크루의 ID
      * @param crewSaveRequestDTO 수정할 크루 내용을 담은 DTO
      * @return 수정된 크루 정보를 담은 DTO
      */
     @Override
-    public CrewResponseDTO updateCrew(Long id, CrewSaveRequestDTO crewSaveRequestDTO) {
+    public CrewResponseDTO updateCrew(AuthInfoDTO currentUser, Long id, CrewSaveRequestDTO crewSaveRequestDTO) {
+        Member currentMember = memberService.findById(currentUser.id());
         Crew crew = crewRepository.findById(id).orElseThrow();
+        CrewMember currentCrewMember = crewMemberRepository.findByCrewAndMember(crew, currentMember).orElseThrow();
+        if (!currentCrewMember.getIsCrewAdmin()) {
+            throw new IllegalArgumentException("Current user is not crew admin");
+        }
         String imageUrl = null;
         if (crewSaveRequestDTO.image() != null) {
             imageUrl = s3ImageUploader.uploadImage(crewSaveRequestDTO.image());
@@ -137,25 +166,31 @@ public class CrewServiceImpl implements CrewService {
 
     /**
      * 삭제하려는 크루 ID를 받아 해당 크루를 DB에서 삭제한다.
-     * @param id 삭제하려는 크루 ID
+     * @param currentUser 현재 로그인한 유저 정보
+     * @param id          삭제하려는 크루 ID
      * @return 삭제된 크루 정보를 담은 DTO
      */
     @Override
-    public CrewResponseDTO deleteCrew(Long id) {
+    public CrewResponseDTO deleteCrew(AuthInfoDTO currentUser, Long id) {
+        Member currentMember = memberService.findById(currentUser.id());
         Crew crew = crewRepository.findById(id).orElseThrow();
+        CrewMember currentCrewMember = crewMemberRepository.findByCrewAndMember(crew, currentMember).orElseThrow();
+        if (!currentCrewMember.getIsCrewAdmin()) {
+            throw new IllegalArgumentException("Current user is not crew admin");
+        }
         crewRepository.delete(crew);
         return CrewResponseDTO.fromEntity(crew);
     }
 
     /**
      * 팔로우하는 사용자의 ID와 팔로우 대상의 ID를 담은 DTO를 받아 팔로우 정보를 DB에 저장한다.
-     * @param memberId 팔로우하는 사용자의 ID
+     * @param currentUser 현재 로그인한 유저 정보
      * @param crewFollowRequestDTO 팔로우 대상의 ID를 담은 DTO
      */
     @Override
-    public void followCrew(Long memberId, CrewFollowRequestDTO crewFollowRequestDTO) {
+    public void followCrew(AuthInfoDTO currentUser, CrewFollowRequestDTO crewFollowRequestDTO) {
         Crew crew = crewRepository.findById(crewFollowRequestDTO.crewId()).orElseThrow();
-        Member member = memberService.findById(memberId);
+        Member member = memberService.findById(currentUser.id());
         followRepository.findByCrewAndMember(crew, member).ifPresentOrElse(follow -> {
             if (follow.getIsFollowed()) {
                 throw new IllegalArgumentException("Crew follower already exists");
@@ -174,13 +209,13 @@ public class CrewServiceImpl implements CrewService {
 
     /**
      * 팔로우하는 사용자의 ID와 팔로우 대상의 ID를 받아 팔로우 정보를 DB에서 삭제한다.
-     * @param memberId 팔로우하는 사용자의 ID
-     * @param crewId 팔로우 대상의 ID
+     * @param currentUser 현재 로그인한 유저 정보
+     * @param crewId      팔로우 대상의 ID
      */
     @Override
-    public void unfollowCrew(Long memberId, Long crewId) {
+    public void unfollowCrew(AuthInfoDTO currentUser, Long crewId) {
         Crew crew = crewRepository.findById(crewId).orElseThrow();
-        Member member = memberService.findById(memberId);
+        Member member = memberService.findById(currentUser.id());
         Follow follow = followRepository.findByCrewAndMember(crew, member).orElseThrow();
         follow.updateIsFollowed(false);
         followRepository.save(follow);
